Split resolver local-scope helpers
