+++ tests/docs_truth/repo_hygiene/typechecker_resolver_validation.rs
@@ -1,6 +1,7 @@
 use super::*;
 
 mod focused_modules;
+mod local_traversal;
 mod support_helpers;
 
 #[test]
@@ -148,33 +149,6 @@ fn typechecker_resolver_type_parameter_helpers_live_in_focused_helper() {
     assert!(
         root.contains("include!(\"resolver_validation_support/resolver_type_parameters.rs\");"),
         "resolver validation support should include focused resolver type-parameter helpers"
-    );
-}
-
-#[test]
-fn typechecker_resolver_pattern_local_traversal_lives_in_focused_helper() {
-    let traversal = read("src/typechecker/resolver_validation/local_traversal.rs");
-    let patterns = read("src/typechecker/resolver_validation/pattern_locals.rs");
-
-    for helper in [
-        "require_resolver_pattern_expr_locals",
-        "require_resolver_pattern_locals",
-        "require_resolver_pattern_binding",
-    ] {
-        assert!(
-            !traversal.contains(&format!("fn {helper}")),
-            "resolver local traversal should not own pattern-local helper: {helper}"
-        );
-        assert!(
-            patterns.contains(&format!("fn {helper}")),
-            "resolver pattern-local traversal should live in focused helper: {helper}"
-        );
-    }
-
-    let root = read("src/typechecker/resolver_validation.rs");
-    assert!(
-        root.contains("include!(\"resolver_validation/pattern_locals.rs\");"),
-        "resolver validation should include focused pattern-local traversal"
     );
 }
